Remove ContentResourceComparator from public API
Since that class is only used by the Annotation class, it seems odd to
have it exposed in the public API (although I may be unaware of the
reason for its place there).
